+++ src/msions/mzml.py
@@ -7,7 +7,7 @@ import pandas as pd
 from typing import List, Union
 
 
-def tic_df(input_mzml: str, level: str = "1", include_ms1_info: bool = False) -> pd.DataFrame:
+def tic_df(input_mzml: str, level: str = "1", include_ms1_info: bool = False, faims: bool = False) -> pd.DataFrame:
 	"""
 	Find the TIC and injection time for each scan in an mzML file.
 	
@@ -20,6 +20,8 @@ def tic_df(input_mzml: str, level: str = "1", include_ms1_info: bool = False) ->
 	include_ms1_info : bool
 		Returns MS1 scan number, m/z, and intensity associated with precursor analyzed in MS2
 		(requires level="2")
+	faims : bool
+		Returns CV associated with each scan.
 		
 	Returns
 	-------
@@ -37,38 +39,84 @@ def tic_df(input_mzml: str, level: str = "1", include_ms1_info: bool = False) ->
 	# create array
 	tic_lst = []
 
-    #record scan, scan time, TIC, & injection time
+    # record scan, scan time, TIC, & injection time
+	# if examining MS1
 	if level == "1":
+		# for each spectrum
 		for spectrum in run:
+			info_lst = []
 			if spectrum.ms_level == 1:
-				tic_lst.append([spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
-								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')])
+				info_lst = [spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
+								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]
+				if faims:
+					info_lst.append(int(float(spectrum.get_element_by_path(['cvParam'])[7].get('value'))))
+				tic_lst.append(info_lst)
+	# if examining MS2
 	elif level == "2":
+		# for each spectrum
 		for spectrum in run:
+			info_lst = []
 			if spectrum.ms_level == 2:
 				if include_ms1_info:
-					tic_lst.append([int(spectrum.get_element_by_path(['precursorList', 'precursor'])[0].get('spectrumRef').split('=')[3]),
+					info_lst = [int(spectrum.get_element_by_path(['precursorList', 'precursor'])[0].get('spectrumRef').split('=')[3]),
 									float(spectrum.get_element_by_path(['precursorList', 'precursor', 
 																		'selectedIonList', 'selectedIon', 'cvParam'])[0].get('value')),
 									float(spectrum.get_element_by_path(['precursorList', 'precursor', 
 																		'selectedIonList', 'selectedIon', 'cvParam'])[2].get('value')),
 									spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
-									spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')])
+									spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]
 				else:
-					tic_lst.append([spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
-									spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')])
+					info_lst = [spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
+								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]
+				if faims:
+					info_lst.append(int(float(spectrum.get_element_by_path(['cvParam'])[7].get('value'))))
+				tic_lst.append(info_lst)
 	elif level == "all":
 		for spectrum in run:
-			tic_lst.append([spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
-							spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')])    
+			info_lst = []
+			if spectrum.ms_level == 2:
+				if include_ms1_info:
+					info_lst = [int(spectrum.get_element_by_path(['precursorList', 'precursor'])[0].get('spectrumRef').split('=')[3]),
+								float(spectrum.get_element_by_path(['precursorList', 'precursor', 
+																	'selectedIonList', 'selectedIon', 'cvParam'])[0].get('value')),
+								float(spectrum.get_element_by_path(['precursorList', 'precursor', 
+																	'selectedIonList', 'selectedIon', 'cvParam'])[2].get('value')),
+								spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
+								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]
+				else:
+					info_lst = [spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
+								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]
+				if faims:
+					info_lst.append(int(float(spectrum.get_element_by_path(['cvParam'])[7].get('value'))))
+			else:
+				if include_ms1_info:
+					info_lst = [-1,
+								-1,
+								-1,
+								spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
+								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]
+				else:
+					info_lst = [spectrum.ID, spectrum.scan_time[0], spectrum.TIC, 
+								spectrum.get_element_by_path(['scanList','scan','cvParam'])[2].get('value')]				
+				if faims:
+					info_lst.append(int(float(spectrum.get_element_by_path(['cvParam'])[7].get('value'))))
+			tic_lst.append(info_lst)
 
 	# create dataframe
-	if level == "2" and include_ms1_info:
-		tic_df = pd.DataFrame(tic_lst,
-							columns=['ms1_scan', 'ms1_mz', 'ms1_int','ms2_scan', 'rt', 'TIC', 'IT'])
+	if (level == "2" or level == "all") and include_ms1_info:
+		if faims:
+			tic_df = pd.DataFrame(tic_lst,
+								  columns=['ms1_scan', 'ms1_mz', 'ms1_int','scan_num', 'rt', 'TIC', 'IT', 'CV'])
+		else:
+			tic_df = pd.DataFrame(tic_lst, 
+								  columns=['ms1_scan', 'ms1_mz', 'ms1_int','scan_num', 'rt', 'TIC', 'IT'])
 	else:
-		tic_df = pd.DataFrame(tic_lst,
-							columns=['scan_num', 'rt', 'TIC', 'IT'])
+		if faims:
+			tic_df = pd.DataFrame(tic_lst,
+								  columns=['scan_num', 'rt', 'TIC', 'IT', 'CV'])	
+		else:	
+			tic_df = pd.DataFrame(tic_lst,
+								  columns=['scan_num', 'rt', 'TIC', 'IT'])
 
 	tic_df['rt'] = tic_df['rt'].round(4)