+++ src/msions/percolator.py
@@ -279,3 +279,86 @@ def id_scans(perc_target, ms2_tic_df):
 
     # create list for whether MS2 was ID'd
 	ms2_tic_df["IDd"] = np.isin(ms2_tic_df["scan_num"], perc_sig['scan'])
+
+
+def match_kro(kro_df: pd.DataFrame, xml_input: pd.DataFrame, ms_input: pd.DataFrame, faims: bool = False):
+	"""
+	Determine if Kronik features were identified or not
+
+	Parameters
+	----------
+	kro_df : pd.DataFrame
+		The pandas DataFrame of Kronik features.
+	xml_input : pd.DataFrame
+		The pandas DataFrame of Percolator XML output.
+	ms_input: pd.DataFrame
+		The pandas DataFrame of MS2 scan and precursor information.
+	faims : bool
+		Whether data is from FAIMS runs
+
+	Examples
+	-------
+	>>> from msions.percolator import match_kro
+	>>> match_kro(kro_df, perc_xml_df, ms_df)
+	""" 
+	# define DataFrames
+	xml_df = xml_input
+	ms_df = ms_input
+
+	# initiate ID list
+	kro_id_lst = len(kro_df)*[0]
+	xml_id_lst = []
+	xml_int_lst = []
+	xml_tic_lst = []
+	xml_it_lst = []
+
+	for row in xml_df.itertuples():
+		# define info to match
+		ms2_ref_scan = row.scan_num
+		ms1_ref_mass = row.exp_mass
+		subset_ms2_df = ms_df[ms_df.scan_num == ms2_ref_scan]
+		ms1_ref_mz = subset_ms2_df.ms1_mz
+		ms1_ref_scan = subset_ms2_df.ms1_scan.reset_index(drop=True)[0]
+
+		# find TIC and IT for MS1
+		subset_ms_df = ms_df[ms_df.scan_num == ms1_ref_scan]
+		xml_tic_lst.append(subset_ms_df.TIC.reset_index(drop=True)[0])
+		xml_it_lst.append(subset_ms_df.IT.reset_index(drop=True)[0])
+		
+		# filter Kronik DataFrame
+		kro_filt = kro_df[(kro_df.first_scan <= ms1_ref_scan) & (kro_df.last_scan >= ms1_ref_scan)]
+		kro_filt = kro_filt[np.isclose(kro_filt.mz, ms1_ref_mz, atol=1.01)]
+		kro_filt = kro_filt[np.isclose(kro_filt.mass, ms1_ref_mass, atol=1.01)]
+
+		# if FAIMS experiment
+		if faims:
+			# define CV to match
+			ref_cv = int(ms_df[ms_df.scan_num == ms2_ref_scan].CV)
+
+			# further filter DataFrame
+			kro_filt = kro_filt[kro_filt.CV == ref_cv]
+
+		# determine indices that are ID'd
+		idx_lst = kro_filt.index
+		num_idxs = len(idx_lst)
+
+		# add ID'd indices to ID lists and feature intensities to XML dataframe
+		if num_idxs == 1:
+			kro_id_lst[idx_lst[0]] = 1
+			xml_int_lst.append(kro_filt.best_int.reset_index(drop=True)[0])
+		elif num_idxs > 1:
+			for idx_single in kro_filt.index:
+				kro_id_lst[idx_single] = num_idxs
+			xml_int_lst.append(max(kro_filt.best_int))
+		else:
+			xml_int_lst.append(0)
+		xml_id_lst.append(num_idxs)
+
+	# add IDs to Kronik DataFrame
+	kro_df["ID_d"] = kro_id_lst
+	xml_df["in_kro"] = xml_id_lst
+	xml_df["best_int"] = xml_int_lst
+	xml_df["TIC"] = xml_tic_lst
+	xml_df["IT"] = xml_it_lst
+	xml_df["ions"] = xml_df['best_int']*xml_df['IT']/1000
+	# xml_df["norm_int"] = xml_df["best_int"]/xml_df["TIC"]