+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 <!--next-version-placeholder-->
 
-## v0.3.2 (2023-02-20)
+## v0.4.0 (2023-03-13)
+
+### Additions
+- Add match_kro function to percolator module
+- Create tests and documentation for match_kro
 
 ### Changes
 - Move find_precursorscan functionality to tic_df function
+- Remove find_precursorscan from documentation
+- Add parameter to find CVs in tic_df function
+- Update simple_df function to allow multiple CVs
 
 ## v0.3.1 (2023-01-15)
 
@@ -17,7 +24,7 @@
 ### Additions
 - Create encyclopedia and utils modules
 - Add find_precursorscan and peak_df to mzml module
-- Add id_scans functions to percolator module
+- Add id_scans function to percolator module
 - Add filter_df and match_rt_mass functions to kronik module
 - Add ability to plot using an mzML file input
 - Create tests and documentation for new functions and modules